# [Enhance] Tweaks on Memory Usage and Performance