--- a/Gemfile
+++ b/Gemfile
@@ -1,4 +1,4 @@
-ruby "2.5.1"
+ruby "2.5.3"
 source 'https://rubygems.org'
 
 git_source(:github) do |repo_name|