--- a/app/models/application_record.rb
+++ b/app/models/application_record.rb
@@ -1,3 +1,0 @@
-class ApplicationRecord < ActiveRecord::Base
-  self.abstract_class = true
-end