--- a/app/workers/reply_worker.rb
+++ b/app/workers/reply_worker.rb
@@ -1,5 +1,3 @@
-require 'line/bot'
-
 class ReplyWorker
   include Sidekiq::Worker
   sidekiq_options retry: false
@@ -17,14 +15,6 @@ class ReplyWorker
   private
 
   def talk(reply)
-    initiate_client
-    @client.reply_message(@token, reply)
-  end
-
-  def initiate_client
-    @client = Line::Bot::Client.new do |config|
-      config.channel_secret = ENV['line_channel_secret']
-      config.channel_token = ENV['line_channel_token']
-    end
+    LineBotClient.reply_message(@token, reply)
   end
 end