--- a/config/application.rb
+++ b/config/application.rb
@@ -2,13 +2,13 @@ require_relative 'boot'
 
 require "rails"
 # Pick the frameworks you want:
-require "active_model/railtie"
+# require "active_model/railtie"
 require "active_job/railtie"
-require "active_record/railtie"
+# require "active_record/railtie"
 require "action_controller/railtie"
 require "action_mailer/railtie"
 require "action_view/railtie"
-require "action_cable/engine"
+# require "action_cable/engine"
 # require "sprockets/railtie"
 # require "rails/test_unit/railtie"