--- a/config/environments/production.rb
+++ b/config/environments/production.rb
@@ -74,5 +74,5 @@ Rails.application.configure do
   end
 
   # Do not dump schema after migrations.
-  config.active_record.dump_schema_after_migration = false
+  # config.active_record.dump_schema_after_migration = false
 end