安安 未來新同事早~
年前看一下 Open Source 來發個 PR 哈哈

## Objective

Little tweaks on Rails memory usage and performance

## Summary

- Upgrade ruby minor version
- Remove unused Rails modules and gems
- Test spec fixed
- Events fixed and optimized

Actually I'm not sure this is a proper fix or not since the original intention in commit message is
![2018-12-30 22 06 04](https://user-images.githubusercontent.com/2749593/50553407-0bffd000-0ce1-11e9-9922-9e36672c9dc5.png)

The old methods actually still loop through all the events actually, so I change to a more proper use of `uniq` with `block` condition
```ruby
# [Old] Same replay token from events still being selected
@uniq_event_tokens = params['events'].map { |e| e['replyToken'] }.uniq
@events = params['events'].select { |e| @uniq_event_tokens.include?(e['replyToken']) }

# [New] Duplicate reply token only selected once and reduce the extra loop
@events = params[:events].uniq { |event| event['replyToken'] }
```

Results:
```
Warming up --------------------------------------
            original     8.408k i/100ms
                 new    14.777k i/100ms
Calculating -------------------------------------
            original     82.067k (± 9.0%) i/s -    411.992k in   5.066976s
                 new    174.802k (± 1.0%) i/s -    886.620k in   5.072660s

Comparison:
                 new:   174802.1 i/s
            original:    82066.6 i/s - 2.13x  slower
```

- Remove the multiple object initializations with LineBot API Client